implementando lógica para o evento estar disponível somente para o usuário que o cadastrou, melhorando algumas funcionalidades do front
