+++ front-vite-react/front-vite-react/src/Components/Content.jsx
@@ -3,7 +3,7 @@ import Login from '../Pages/Login';
 import Register from '../Pages/Register';
 import Home from '../Pages/Home';
 import Calendar from "../Pages/Calendar";
-import Event from "../Pages/Event";
+import NewEvent from "../Pages/NewEvent";
 import Logout from "../Pages/Logout";
 import User from "../Pages/User";
 import ShowEvent from "../Pages/ShowEvent";
@@ -18,7 +18,7 @@ const AppRoutes = () => {
                     <Route path="/login" element={<Login />} />
                     <Route path="/register" element={<Register />} />
                     <Route path="/calendar" element={<Calendar />} />
-                    <Route path="/event" element={<Event />} />
+                    <Route path="/event" element={<NewEvent />} />
                     <Route path="/user" element={<User />} />
                     <Route path="/event/:id" element={<ShowEvent />} />
                     <Route path="/logout" element={<Logout />} />
