--- front-vite-react/front-vite-react/src/Pages/EditEvent.jsx
@@ -1,11 +1,0 @@
-const EditEvent = () =>{
-
-    return(
-        <div>
-            <h1>evento</h1>
-        </div>
-    );
-
-}
-
-export default EditEvent;
